Add files via upload
This is testing development world SCREENS from start of this Month (October 2025). My experience allows me to store previous versions of the process of development with some functionalities "new" at the stage of the process in that time. 

Current version which I'm working on have more things implemented like:
- Parkours
- better optimization of scripts
- Phone functionality
- More NPCs on world
- etc.

But I will share it with later via Upwork, so our contract will be fullfilled